+++ test-workspace/config/spa/extract-css.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   entry: {
     'css-a': './spa/scripts/css-page-a.js',
     'css-b': './spa/scripts/css-page-b.js'