+++ test-workspace/config/spa/main-sync.js
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
   entry: {
     // lib: ['jquery'], // for test
     main: './spa/scripts/main-sync.js'