+++ test-workspace/config/spa/vendor-all.js
@@ -1,6 +1,6 @@
-export default {
+module.exports = {
   entry: {
     main: './spa/scripts/vendor-main.js'
   },
-  extractAllVendors: true
+  splitAllVendors: true
 };