+++ test-workspace/php/resources/views/index.blade.php
@@ -23,6 +23,7 @@ use \App\Helpers as helpers;
       <span>Hello</span>
       <img src="<?php echo helpers\balm('img/logo.svg', 'web'); ?>" alt="BalmJS" />
       <a href="//balmjs.com/">BalmJS</a>
+      <span>for <b>PHP</b></span>
     </div>
     <script src="<?php echo helpers\balm('js/app.js', 'web'); ?>"></script>
   </body>