+++ test-workspace/spa/scripts/vendor-main.js
@@ -1,2 +1,2 @@
-import fooA from './vendor-a';
-import fooB from './vendor-b';
+import './vendor-a';
+import './vendor-b';